refactor: move mdx components into folder
