+++ src/all.ts
@@ -40,26 +40,26 @@ import './components/lightbox/LightBox';
 import './components/list-item/ListItem';
 import './components/list/CollapsibleList';
 import './components/list/List';
-import './components/mdx-anchor/MdxAnchor';
-import './components/mdx-badge/MdxBadge';
-import './components/mdx-banner/MdxBanner';
-import './components/mdx-bread-crumbs/MdxBreadCrumbs';
-import './components/mdx-button/MdxButton';
-import './components/mdx-card-grid/MdxCardGrid';
-import './components/mdx-card/MdxCard';
-import './components/mdx-code-block/MdxCodeBlock';
-import './components/mdx-collapse/MdxCollapse';
-import './components/mdx-copy-value/MdxCopyValue';
-import './components/mdx-hero/MdxHero';
-import './components/mdx-image/MdxImage';
-import './components/mdx-progress-checkbox/MdxProgressCheckbox';
-import './components/mdx-related-content/MdxRelatedContent';
-import './components/mdx-section/MdxSection';
-import './components/mdx-subtitle/MdxSubtitle';
-import './components/mdx-table/MdxTable';
-import './components/mdx-tabs/MdxTab';
-import './components/mdx-tabs/MdxTabs';
-import './components/mdx-video-guide/MdxVideoGuide';
+import './components/mdx/mdx/mdx-anchor/MdxAnchor';
+import './components/mdx/mdx-badge/MdxBadge';
+import './components/mdx/mdx-banner/MdxBanner';
+import './components/mdx/mdx-bread-crumbs/MdxBreadCrumbs';
+import './components/mdx/mdx-button/MdxButton';
+import './components/mdx/mdx-card-grid/MdxCardGrid';
+import './components/mdx/mdx-card/MdxCard';
+import './components/mdx/mdx-code-block/MdxCodeBlock';
+import './components/mdx/mdx-collapse/MdxCollapse';
+import './components/mdx/mdx-copy-value/MdxCopyValue';
+import './components/mdx/mdx-hero/MdxHero';
+import './components/mdx/mdx-image/MdxImage';
+import './components/mdx/mdx-progress-checkbox/MdxProgressCheckbox';
+import './components/mdx/mdx-related-content/MdxRelatedContent';
+import './components/mdx/mdx-section/MdxSection';
+import './components/mdx/mdx-subtitle/MdxSubtitle';
+import './components/mdx/mdx-table/MdxTable';
+import './components/mdx/mdx-tabs/MdxTab';
+import './components/mdx/mdx-tabs/MdxTabs';
+import './components/mdx/mdx-video-guide/MdxVideoGuide';
 import './components/menu/Menu';
 import './components/menu/MenuItem';
 import './components/nested-data-table/NestedDataTable';
