+++ src/components/mdx/mdx-collapse/MdxCollapse.tsx
