+++ src/components/mdx/mdx-image/MdxImage.tsx
