+++ src/components/mdx/mdx-table/MdxTable.tsx
